Pull chart text from Google Sheet and format it

This function retrieves chart-related text from a Google Sheet and formats it. It assumes that the spreadsheet contains the following columns: viz_id, title, subtitle, source, note, x, and y.

Usage

get_chart_text_from_gsheet(
  sheet_id,
  viz_unique_id,
  title_wrap = 150,
  subtitle_wrap = 150,
  caption_wrap = 80
)

Arguments

sheet_id

Character. The Google Sheets ID.

viz_unique_id

Character. The unique visualization ID in the sheet.

title_wrap

Integer. Character width for wrapping the title. Default is 150.

subtitle_wrap

Integer. Character width for wrapping the subtitle. Default is 150.

caption_wrap

Integer. Character width for wrapping the source and notes. Default is 80.

Value

A named list with elements: title, subtitle (if available), x (if available), y (if available), and caption (if available).